Move show_invalid_search_error to utils
